Adding error module
